requirements.txt
flask
google-generativeai
openai>=1.0.0
requests
pandas
sqlalchemy
unidecode
numpy
matplotlib
requests
beautifulsoup4
lxml
psycopg2-binary
python-dotenv
Flask
gunicorn
google-generativeai
cloud-sql-python-connector
argparse 
